SPI Split interface refactor.
