+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/FileScanNode.java
@@ -27,6 +27,8 @@ import org.apache.doris.catalog.TableIf;
 import org.apache.doris.common.UserException;
 import org.apache.doris.planner.FileLoadScanNode;
 import org.apache.doris.planner.PlanNodeId;
+import org.apache.doris.qe.ConnectContext;
+import org.apache.doris.spi.Split;
 import org.apache.doris.statistics.StatisticalType;
 import org.apache.doris.thrift.TExplainLevel;
 import org.apache.doris.thrift.TExpr;
@@ -41,9 +43,12 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
+import org.apache.hadoop.fs.BlockLocation;
+import org.apache.hadoop.fs.Path;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.io.IOException;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
@@ -54,6 +59,8 @@ import java.util.Map;
  */
 public class FileScanNode extends ExternalScanNode {
     private static final Logger LOG = LogManager.getLogger(FileScanNode.class);
+
+    public static final long DEFAULT_SPLIT_SIZE = 128 * 1024 * 1024; // 128MB
 
     // For explain
     protected long inputSplitsNum = 0;
@@ -152,24 +159,17 @@ public class FileScanNode extends ExternalScanNode {
         return output.toString();
     }
 
-    // TODO: Keep 2 versions of createScanRangeLocations, will fix this while refactor split and assignment code.
+    // TODO: This api is for load job only. Will remove it later.
     protected void createScanRangeLocations(FileLoadScanNode.ParamCreateContext context,
-                                            FileScanProviderIf scanProvider)
+                                            LoadScanProvider scanProvider)
             throws UserException {
         scanProvider.createScanRangeLocations(context, backendPolicy, scanRangeLocations);
     }
 
-    protected void createScanRangeLocations(List<Expr> conjuncts, TFileScanRangeParams params,
-                                            FileScanProviderIf scanProvider)
-            throws UserException {
-        scanProvider.createScanRangeLocations(conjuncts, params, backendPolicy, scanRangeLocations);
-    }
-
-    protected void setDefaultValueExprs(FileScanProviderIf scanProvider,
+    protected void setDefaultValueExprs(TableIf tbl,
                                         Map<String, SlotDescriptor> slotDescByName,
                                         TFileScanRangeParams params,
                                         boolean useVarcharAsNull) throws UserException {
-        TableIf tbl = scanProvider.getTargetTable();
         Preconditions.checkNotNull(tbl);
         TExpr tExpr = new TExpr();
         tExpr.setNodes(Lists.newArrayList());
@@ -209,5 +209,55 @@ public class FileScanNode extends ExternalScanNode {
                 }
             }
         }
+    }
+
+    protected List<Split> splitFile(Path path, long blockSize, BlockLocation[] blockLocations, long length,
+                                  boolean splittable, List<String> partitionValues) throws IOException {
+        if (blockLocations == null) {
+            blockLocations = new BlockLocation[0];
+        }
+        long splitSize = ConnectContext.get().getSessionVariable().getFileSplitSize();
+        if (splitSize <= 0) {
+            splitSize = blockSize;
+        }
+        // Min split size is DEFAULT_SPLIT_SIZE(128MB).
+        splitSize = Math.max(splitSize, DEFAULT_SPLIT_SIZE);
+        List<Split> result = Lists.newArrayList();
+        if (!splittable) {
+            LOG.debug("Path {} is not splittable.", path);
+            String[] hosts = blockLocations.length == 0 ? null : blockLocations[0].getHosts();
+            result.add(new FileSplit(path, 0, length, length, hosts, partitionValues));
+            return result;
+        }
+        long bytesRemaining;
+        for (bytesRemaining = length; (double) bytesRemaining / (double) splitSize > 1.1D;
+                bytesRemaining -= splitSize) {
+            int location = getBlockIndex(blockLocations, length - bytesRemaining);
+            String[] hosts = location == -1 ? null : blockLocations[location].getHosts();
+            result.add(new FileSplit(path, length - bytesRemaining, splitSize, length, hosts, partitionValues));
+        }
+        if (bytesRemaining != 0L) {
+            int location = getBlockIndex(blockLocations, length - bytesRemaining);
+            String[] hosts = location == -1 ? null : blockLocations[location].getHosts();
+            result.add(new FileSplit(path, length - bytesRemaining, bytesRemaining, length, hosts, partitionValues));
+        }
+
+        LOG.debug("Path {} includes {} splits.", path, result.size());
+        return result;
+    }
+
+    private int getBlockIndex(BlockLocation[] blkLocations, long offset) {
+        if (blkLocations == null || blkLocations.length == 0) {
+            return -1;
+        }
+        for (int i = 0; i < blkLocations.length; ++i) {
+            if (blkLocations[i].getOffset() <= offset
+                    && offset < blkLocations[i].getOffset() + blkLocations[i].getLength()) {
+                return i;
+            }
+        }
+        BlockLocation last = blkLocations[blkLocations.length - 1];
+        long fileLength = last.getOffset() + last.getLength() - 1L;
+        throw new IllegalArgumentException(String.format("Offset %d is outside of file (0..%d)", offset, fileLength));
     }
 }
