+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/LoadScanProvider.java
@@ -18,7 +18,6 @@
 package org.apache.doris.planner.external;
 
 import org.apache.doris.analysis.Analyzer;
-import org.apache.doris.analysis.Expr;
 import org.apache.doris.analysis.ImportColumnDesc;
 import org.apache.doris.analysis.IntLiteral;
 import org.apache.doris.analysis.SlotRef;
@@ -55,7 +54,7 @@ import com.google.common.collect.Maps;
 import java.util.List;
 import java.util.Map;
 
-public class LoadScanProvider implements FileScanProviderIf {
+public class LoadScanProvider {
 
     private FileGroupInfo fileGroupInfo;
     private TupleDescriptor destTupleDesc;
@@ -65,27 +64,22 @@ public class LoadScanProvider implements FileScanProviderIf {
         this.destTupleDesc = destTupleDesc;
     }
 
-    @Override
     public TFileFormatType getFileFormatType() throws DdlException, MetaNotFoundException {
         return null;
     }
 
-    @Override
     public TFileType getLocationType() throws DdlException, MetaNotFoundException {
         return null;
     }
 
-    @Override
     public Map<String, String> getLocationProperties() throws MetaNotFoundException, DdlException {
         return null;
     }
 
-    @Override
     public List<String> getPathPartitionKeys() throws DdlException, MetaNotFoundException {
         return null;
     }
 
-    @Override
     public FileLoadScanNode.ParamCreateContext createContext(Analyzer analyzer) throws UserException {
         FileLoadScanNode.ParamCreateContext ctx = new FileLoadScanNode.ParamCreateContext();
         ctx.destTupleDescriptor = destTupleDesc;
@@ -138,7 +132,6 @@ public class LoadScanProvider implements FileScanProviderIf {
         return "";
     }
 
-    @Override
     public void createScanRangeLocations(FileLoadScanNode.ParamCreateContext context,
                                          FederationBackendPolicy backendPolicy,
                                          List<TScanRangeLocations> scanRangeLocations) throws UserException {
@@ -147,18 +140,10 @@ public class LoadScanProvider implements FileScanProviderIf {
         fileGroupInfo.createScanRangeLocations(context, backendPolicy, scanRangeLocations);
     }
 
-    @Override
-    public void createScanRangeLocations(List<Expr> conjuncts, TFileScanRangeParams params,
-                                         FederationBackendPolicy backendPolicy,
-                                         List<TScanRangeLocations> scanRangeLocations) throws UserException {
-    }
-
-    @Override
     public int getInputSplitNum() {
         return fileGroupInfo.getFileStatuses().size();
     }
 
-    @Override
     public long getInputFileSize() {
         long res = 0;
         for (TBrokerFileStatus fileStatus : fileGroupInfo.getFileStatuses()) {
@@ -250,7 +235,6 @@ public class LoadScanProvider implements FileScanProviderIf {
         }
     }
 
-    @Override
     public TableIf getTargetTable() {
         return fileGroupInfo.getTargetTable();
     }
