+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/iceberg/IcebergHMSSource.java
@@ -26,8 +26,9 @@ import org.apache.doris.common.MetaNotFoundException;
 import org.apache.doris.common.UserException;
 import org.apache.doris.datasource.ExternalCatalog;
 import org.apache.doris.planner.ColumnRange;
-import org.apache.doris.planner.external.HiveScanProvider;
+import org.apache.doris.planner.external.HiveScanNode;
 import org.apache.doris.thrift.TFileAttributes;
+import org.apache.doris.thrift.TFileTextScanRangeParams;
 
 import org.apache.iceberg.TableProperties;
 
@@ -36,15 +37,14 @@ import java.util.Map;
 public class IcebergHMSSource implements IcebergSource {
 
     private final HMSExternalTable hmsTable;
-    private final HiveScanProvider hiveScanProvider;
-
     private final TupleDescriptor desc;
+    private final Map<String, ColumnRange> columnNameToRange;
 
     public IcebergHMSSource(HMSExternalTable hmsTable, TupleDescriptor desc,
                             Map<String, ColumnRange> columnNameToRange) {
-        this.hiveScanProvider = new HiveScanProvider(hmsTable, desc, columnNameToRange);
         this.hmsTable = hmsTable;
         this.desc = desc;
+        this.columnNameToRange = columnNameToRange;
     }
 
     @Override
@@ -54,8 +54,8 @@ public class IcebergHMSSource implements IcebergSource {
 
     @Override
     public String getFileFormat() throws DdlException, MetaNotFoundException {
-        return hiveScanProvider.getRemoteHiveTable().getParameters()
-                .getOrDefault(TableProperties.DEFAULT_FILE_FORMAT, TableProperties.DEFAULT_FILE_FORMAT_DEFAULT);
+        return hmsTable.getRemoteTable().getParameters()
+            .getOrDefault(TableProperties.DEFAULT_FILE_FORMAT, TableProperties.DEFAULT_FILE_FORMAT_DEFAULT);
     }
 
     public org.apache.iceberg.Table getIcebergTable() throws MetaNotFoundException {
@@ -64,12 +64,19 @@ public class IcebergHMSSource implements IcebergSource {
 
     @Override
     public TableIf getTargetTable() {
-        return hiveScanProvider.getTargetTable();
+        return hmsTable;
     }
 
     @Override
     public TFileAttributes getFileAttributes() throws UserException {
-        return hiveScanProvider.getFileAttributes();
+        TFileTextScanRangeParams textParams = new TFileTextScanRangeParams();
+        textParams.setColumnSeparator(hmsTable.getRemoteTable().getSd().getSerdeInfo().getParameters()
+                .getOrDefault(HiveScanNode.PROP_FIELD_DELIMITER, HiveScanNode.DEFAULT_FIELD_DELIMITER));
+        textParams.setLineDelimiter(HiveScanNode.DEFAULT_LINE_DELIMITER);
+        TFileAttributes fileAttributes = new TFileAttributes();
+        fileAttributes.setTextParams(textParams);
+        fileAttributes.setHeaderType("");
+        return fileAttributes;
     }
 
     @Override
